:recycle: Update editor files
