+++ package.json
@@ -41,6 +41,7 @@
     "@types/react": "^17",
     "@types/react-dom": "^17",
     "@types/sass": "^1",
+    "@types/sortablejs": "^1",
     "@types/string-similarity": "^4",
     "@types/uuid": "^8.3.0",
     "@types/webpack": "^4.41.25",
@@ -100,9 +101,11 @@
     "react-player": "^2.8.2",
     "react-query": "^2.26.4",
     "react-query-devtools": "^2.6.3",
+    "react-sortablejs": "^6.0.0",
     "resize-observer-polyfill": "^1.5.1",
     "safe-buffer": "^5.2.1",
     "sass": "^1.32.2",
+    "sortablejs": "^1.13.0",
     "string-similarity": "^4.0.4",
     "typeorm": "^0.2.29",
     "uuid": "^8.3.2",
