+++ src/components/activityEditor/index.ts
@@ -1,0 +1,7 @@
+import dynamic from "next/dynamic";
+
+import { SimpleActivityPreview } from "./SimpleActivityPreview";
+
+const SimpleActivityEditor = dynamic(() => import("./SimpleActivityEditor"), { ssr: false });
+
+export { SimpleActivityEditor, SimpleActivityPreview };
