+++ src/components/activityEditor/previews/ImagePreview.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import type { PreviewProps } from "./editing.types";
+import type { PreviewProps } from "../editing.types";
 
 export const ImagePreview: React.FC<PreviewProps> = ({ value }: PreviewProps) => {
   return (
