+++ src/components/activityEditor/previews/TextPreview.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import type { PreviewProps } from "./editing.types";
+import type { PreviewProps } from "../editing.types";
 
 export const TextPreview: React.FC<PreviewProps> = ({ value }: PreviewProps) => {
   return (
