--- src/components/editing/ImageEditor/index.ts
@@ -1,3 +1,0 @@
-import dynamic from "next/dynamic";
-
-export const ImageEditor = dynamic(() => import("./ImageEditor"), { ssr: false });
