--- src/components/editing/TextEditor/index.ts
@@ -1,3 +1,0 @@
-import dynamic from "next/dynamic";
-
-export const TextEditor = dynamic(() => import("./TextEditor"), { ssr: false });
