--- src/components/editing/VideoEditor/index.ts
@@ -1,3 +1,0 @@
-import dynamic from "next/dynamic";
-
-export const VideoEditor = dynamic(() => import("./VideoEditor"), { ssr: false });
