+++ src/pages/se-presenter/thematique/3.tsx
@@ -5,10 +5,8 @@ import { Button } from "@material-ui/core";
 
 import { Base } from "src/components/Base";
 import { Steps } from "src/components/Steps";
+import { SimpleActivityPreview } from "src/components/activityEditor";
 import { BackButton } from "src/components/buttons/BackButton";
-import { ImagePreview } from "src/components/editing/ImagePreview";
-import { TextPreview } from "src/components/editing/TextPreview";
-import { VideoPreview } from "src/components/editing/VideoPreview";
 import { ActivityContext } from "src/contexts/activityContext";
 
 const PresentationStep3: React.FC = () => {
@@ -43,18 +41,7 @@ const PresentationStep3: React.FC = () => {
               Publier
             </Button>
           </div>
-          {activity.processedContent.map((p) => {
-            if (p.type === "text") {
-              return <TextPreview value={p.value} key={p.id} />;
-            }
-            if (p.type === "image") {
-              return <ImagePreview value={p.value} key={p.id} />;
-            }
-            if (p.type === "video") {
-              return <VideoPreview value={p.value} key={p.id} />;
-            }
-            return null;
-          })}
+          <SimpleActivityPreview />
         </div>
       </div>
     </Base>
